:recycle: move to workout level aspects
