--- packages/curriculum-ast-validators/__tests__/fixtures/insight/invalid-yaml-aspects.md
@@ -1,12 +1,0 @@
----
-author: enki
-aspects:
-  - invalid
----
-
-# Headline
-
----
-## Content
-
-text
