+++ packages/curriculum-ast-validators/__tests__/fixtures/insight/invalid-yaml-category.md
@@ -1,7 +1,5 @@
 ---
 author: enki
-aspects:
-  - introduction
 category: invalid
 ---
