+++ packages/curriculum-ast-validators/insight/yaml.js
@@ -1,9 +1,6 @@
 const visit = require('unist-util-visit');
 const jsYaml = require('js-yaml');
-const {
-  insightAspects,
-  insightCategories,
-} = require('@enkidevs/curriculum-helpers');
+const { insightCategories } = require('@enkidevs/curriculum-helpers');
 
 module.exports = function validateYaml(ast) {
   let yamlNodeCount = 0;
@@ -21,18 +18,6 @@ module.exports = function validateYaml(ast) {
   }
 
   const parsedValue = jsYaml.safeLoad(yamlNode.value, 'utf8');
-
-  const invalidAspects = parsedValue.aspects.filter(
-    (aspect) => !insightAspects.includes(aspect)
-  );
-
-  if (invalidAspects.length > 0) {
-    throw new Error(
-      `Invalid insight aspects "${invalidAspects.join(
-        '", "'
-      )}". Valid aspects are "${insightAspects.join('", "')}"`
-    );
-  }
 
   if (!insightCategories.includes(parsedValue.category)) {
     throw new Error(
