+++ packages/curriculum-helpers/lib/index.js
@@ -3,7 +3,7 @@ const compactAst = require('./compact-ast');
 const sectionNames = require('./section-names');
 const questionTypes = require('./question-types');
 const insightCategories = require('./insight-categories');
-const insightAspects = require('./insight-aspects');
+const workoutAspects = require('./workout-aspects');
 const insightLinkResourceTypes = require('./insight-link-resource-types');
 const internalLinks = require('./internal-links');
 
@@ -14,6 +14,6 @@ module.exports = {
   questionTypes,
   insightCategories,
   insightLinkResourceTypes,
-  insightAspects,
+  workoutAspects,
   internalLinks,
 };
