+++ packages/curriculum-helpers/lib/workout-aspects.js
